+++ src/ska_tmc_common/test_helpers/helper_mccs_master_leaf_node_device.py
@@ -24,7 +24,7 @@ class HelperMCCSMasterLeafNode(HelperBaseDevice):
     def init_device(self) -> None:
         super().init_device()
         self.dev_name = self.get_name()
-        self._isSubsystemAvailable = False
+        self._isSubsystemAvailable = True
         self._delay: int = 2
         self._raise_exception = False