+++ CHANGELOG.rst
@@ -25,6 +25,10 @@ Added
 * EndScan command has been added in HelperDishLNDevice
 * **scanID** attribute has been introduced in HelperDishDevice
 
+[0.15.2]
+************
+* Updated device availability to be **True** by default
+
 
 Fixed
 -----
@@ -34,6 +38,11 @@ Fixed
 
 * Updated **push_command_result** method from the HelperBaseDevice to take correct number of arguments
 * The sequence of executing cleanup and **update_task_status** method is reversed in the Tracker Thread
+
+[0.15.2]
+************
+
+* Updated **HelperDishDevice** to add EndScan command to reset **scanID** attribute.
 
 [0.14.0]
 ************